+++ PocoGenerator/PocoGenerator.Domain.Models/BaseObjects/SysTypes.cs
@@ -9,7 +9,7 @@ namespace PocoGenerator.Domain.Models.BaseObjects
     public class SysTypes
     {
         public string name { get; set; }
-        public Int16 xtype { get; set; }
+        public byte xtype { get; set; }
 
         public virtual SysColumns Column { get; set; }
     }
